requirements.txt
pandas==1.0.5
numpy==1.19.5
python-decouple==3.4
sklearn==0.23.1
pandas-datareader==0.9.0
tensorflow==2.5.0
matplotlib==3.2.2
